requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dj-config-url==0.1.1
Django==4.1.6
django-appconf==1.0.5
django-braces==1.15.0
django-embed-video==1.4.8
django-memcache-status==2.2
django-staticfiles==1.2.1
django-yugabytedb==4.0.0.post1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
python-memcached==1.59
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
uWSGI==2.0.21
whitenoise==6.4.0
